requirements.txt
bidict==0.23.1
blinker==1.9.0
click==8.2.1
colorama==0.4.6
Flask==3.1.1
flask-cors==6.0.0
Flask-SocketIO==5.5.1
h11==0.16.0
itsdangerous==2.2.0
Jinja2==3.1.6
keyboard==0.13.5
MarkupSafe==3.0.2
MouseInfo==0.1.3
pillow==11.2.1
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyJWT==2.10.1
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dotenv==1.1.0
python-engineio==4.12.2
python-socketio==5.13.0
pytweening==1.2.0
qrcode==8.2
simple-websocket==1.1.0
Werkzeug==3.1.3
wsproto==1.2.0
